Ajout confirm delete quiz + affichage nb question + refonte page gestion

+++ frontend/src/app/quizzes/quiz-list/quiz-list.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { QuizService } from '../../../services/quiz.service';
 import { Quiz } from '../../../models/quiz.model';
+import {MatDialog} from '@angular/material/dialog';
+import {DialogDeleteQuizComponent} from '../dialog-delete-quiz/dialog-delete-quiz.component';
 
 @Component({
   selector: 'app-quiz-list',
@@ -11,7 +13,7 @@ export class QuizListComponent implements OnInit {
 
   public quizList: Quiz[] = [];
 
-  constructor(public quizService: QuizService) {
+  constructor(public quizService: QuizService, public dialog: MatDialog) {
     this.quizService.quizzes$.subscribe((quiz) => this.quizList = quiz);
   }
 
@@ -19,7 +21,16 @@ export class QuizListComponent implements OnInit {
   }
 
   deleteQuiz(quiz: Quiz) {
-    console.log('Delete ', quiz);
-    this.quizService.deleteQuiz(quiz);
+    const dialogRef = this.dialog.open(DialogDeleteQuizComponent, {
+      width: '450px',
+      height: '250px'
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+        if ( result ) {
+            console.log('Delete ', quiz);
+            this.quizService.deleteQuiz(quiz);
+        }
+    });
   }
 }
